+++ build-logic/plugins/convention/src/main/kotlin/com/flipperdevices/buildlogic/SourceInstall.kt
@@ -1,0 +1,7 @@
+package com.flipperdevices.buildlogic
+
+enum class SourceInstall {
+    GITHUB,
+    GOOGLE_PLAY,
+    UNKNOWN
+}